Polish 'Document methods of starting Testcontainer containers'
See gh-44187
