+++ spring-boot-project/spring-boot-docs/src/main/kotlin/org/springframework/boot/docs/testing/testcontainers/dynamicproperties/MyIntegrationTests.kt
@@ -29,7 +29,7 @@ class MyIntegrationTests {
 
 	@Test
 	fun myTest() {
-		// ...
+		/**/ println()
 	}
 
 	companion object {
