+++ spring-boot-project/spring-boot-docs/src/main/kotlin/org/springframework/boot/docs/testing/testcontainers/junitextension/MyIntegrationTests.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2025 the original author or authors.
+ * Copyright 2012-2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,25 +14,30 @@
  * limitations under the License.
  */
 
-package org.springframework.boot.docs.testing.testcontainers.beandeclaration;
+package org.springframework.boot.docs.testing.testcontainers.junitextension
 
 import org.junit.jupiter.api.Test;
-import org.testcontainers.containers.MongoDBContainer;
+import org.testcontainers.containers.Neo4jContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
 
+@Testcontainers
 @SpringBootTest
-@Import(BeanDeclarationConfig.class)
-class SpringTest {
-
-	@Autowired
-	private MongoDBContainer mongo;
+class MyIntegrationTests {
 
 	@Test
-	void doTest() {
-		System.out.println("Mongo db is running: " + this.mongo.isRunning());
+	fun myTest() {
+		/**/ println()
 	}
 
+	companion object {
+
+		@Container
+		@JvmStatic
+		val neo4j = Neo4jContainer("neo4j:5");
+
+	}
 }
+
