+++ spring-boot-project/spring-boot-docs/src/main/kotlin/org/springframework/boot/docs/testing/testcontainers/serviceconnections/MyIntegrationTests.kt
@@ -30,7 +30,7 @@ class MyIntegrationTests {
 
 	@Test
 	fun myTest() {
-		// ...
+		/**/ println()
 	}
 
 	companion object {
