+++ spring-boot-project/spring-boot-docs/src/main/kotlin/org/springframework/boot/docs/testing/testcontainers/springbeans/MyTestConfiguration.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2024 the original author or authors.
+ * Copyright 2012-2025 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,26 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.springframework.boot.docs.testing.testcontainers.springbeans
 
-package org.springframework.boot.docs.testing.testcontainers.vanilla;
+import org.springframework.boot.test.context.TestConfiguration
+import org.springframework.context.annotation.Bean
+import org.testcontainers.containers.MongoDBContainer
+import org.testcontainers.utility.DockerImageName
 
-import org.junit.jupiter.api.Test;
-import org.testcontainers.containers.Neo4jContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
+@TestConfiguration(proxyBeanMethods = false)
+class MyTestConfiguration {
 
-import org.springframework.boot.test.context.SpringBootTest;
-
-@Testcontainers
-@SpringBootTest
-class MyIntegrationTests {
-
-	@Container
-	static Neo4jContainer<?> neo4j = new Neo4jContainer<>("neo4j:5");
-
-	@Test
-	void myTest() {
-		// ...
+	@Bean
+	fun mongoDbContainer(): MongoDBContainer {
+		return MongoDBContainer(DockerImageName.parse("mongo:5.0"))
 	}
 
 }
